# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Viral Clip AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/viral-clip/clips/{clip_id}/exports": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "Create Export",
        "operationId": "create_export_api_viral_clip_clips__clip_id__exports_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportWithJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "List Exports",
        "operationId": "list_exports_api_viral_clip_clips__clip_id__exports_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExportOut"
                  },
                  "title": "Response List Exports Api Viral Clip Clips  Clip Id  Exports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports/{export_id}": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Get Export",
        "operationId": "get_export_api_exports__export_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports/{export_id}/download": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Download Export",
        "operationId": "download_export_api_exports__export_id__download_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/viral-clip/clip-batches/{batch_id}/exports": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "List Exports For Batch",
        "operationId": "list_exports_for_batch_api_viral_clip_clip_batches__batch_id__exports_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ExportOut"
                    }
                  },
                  "title": "Response List Exports For Batch Api Viral Clip Clip Batches  Batch Id  Exports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportCreate": {
        "properties": {
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "1080p"
          },
          "fps": {
            "type": "integer",
            "title": "Fps",
            "default": 30
          },
          "aspect_ratio": {
            "type": "string",
            "title": "Aspect Ratio",
            "default": "9:16"
          },
          "use_brand_kit": {
            "type": "boolean",
            "title": "Use Brand Kit",
            "default": true
          },
          "use_ai_dub": {
            "type": "boolean",
            "title": "Use Ai Dub",
            "default": true
          }
        },
        "type": "object",
        "title": "ExportCreate"
      },
      "ExportOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "clip_id": {
            "type": "integer",
            "title": "Clip Id"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution"
          },
          "fps": {
            "type": "integer",
            "title": "Fps"
          },
          "aspect_ratio": {
            "type": "string",
            "title": "Aspect Ratio"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "clip_id",
          "resolution",
          "fps",
          "aspect_ratio",
          "status",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "ExportOut"
      },
      "ExportWithJob": {
        "properties": {
          "export": {
            "$ref": "#/components/schemas/ExportOut"
          },
          "job": {
            "$ref": "#/components/schemas/ProcessingJobOut"
          }
        },
        "type": "object",
        "required": [
          "export",
          "job"
        ],
        "title": "ExportWithJob"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProcessingJobOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "video_source_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Source Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "result_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Summary"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "ProcessingJobOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}